requirements.txt
google-api-python-client==2.137.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-pubsub==2.21.5
google-cloud-firestore==2.16.0
python-dotenv==1.0.1

